+++ visualization/app/codeCharta/stores/sharedView/store/metricRules/metricRules.reducer.ts
@@ -1,6 +1,7 @@
 import { createReducer, on } from "@ngrx/store"
 import { MetricRule } from "../../../../model/codeCharta.model"
 import { setState } from "../../../../util/setState.reducer.factory"
+import { clearRulesOfType } from "../rules/rules.actions"
 import { addMetricRule, removeMetricRule, setMetricRules } from "./metricRules.actions"
 
 export const defaultMetricRules: MetricRule[] = []
@@ -8,5 +9,6 @@ export const metricRules = createReducer(
     defaultMetricRules,
     on(setMetricRules, setState(defaultMetricRules)),
     on(addMetricRule, (state, action) => (state.some(rule => rule.id === action.rule.id) ? state : [...state, action.rule])),
-    on(removeMetricRule, (state, action) => state.filter(rule => rule.id !== action.id))
+    on(removeMetricRule, (state, action) => state.filter(rule => rule.id !== action.id)),
+    on(clearRulesOfType, (state, action) => state.filter(rule => rule.type !== action.blacklistType))
 )